処理共通化
